basic flashes for login

+++ application.py
@@ -39,7 +39,7 @@ def showCategory(name):
 
 def addCategory(name):
     try:
-        new_category = Category(name=name)
+        new_category = Category(name=name,user=getCurrentUser())
         session.add(new_category)
         session.commit()
         return True
@@ -63,7 +63,7 @@ def showItem(categoryName,itemName):
 def newItem(name,description,categoryName):
     try:
         category = showCategory(categoryName)
-        new_item = Item(name=name,description=description,category=category)
+        new_item = Item(name=name,description=description,category=category,user=getCurrentUser())
         session.add(new_item)
         session.commit()
         return True
@@ -96,6 +96,13 @@ def isLoggedIn():
 def getUserID(id):
     user = session.query(User).filter_by(id=id).first()
     return user
+
+def getCurrentUser():
+    try:
+        user = session.query(User).filter_by(id=login_session['user_id']).first()
+        return user
+    except:
+        return False
 
 def createUser(user):
     try:
@@ -142,8 +149,7 @@ def categoryDisplay(categoryName):
 @app.route('/catalog/<categoryName>/new',methods = ['GET','POST'])
 def newCategoryItem(categoryName):
     if not isLoggedIn():
-        return redirect('/login')
-    if not isLoggedIn():
+        flash('login is required to create a new item','warning')
         return redirect('/login')
     if request.method == 'POST':
         try:
@@ -175,6 +181,7 @@ def newCategoryItem(categoryName):
 @app.route('/catalog/new', methods = ['GET','POST'])
 def newCategory():
     if not isLoggedIn():
+        flash('login is required to create a new category','warning')
         return redirect('/login')
     if request.method == 'POST':
         try:
@@ -204,10 +211,12 @@ def categoryItem(categoryName,itemName):
 #@auth.login_required
 def categoryItemEdit(categoryName,itemName):
     if not isLoggedIn():
+        flash('login is required to create a new category','warning')
         return redirect('/login')
-    creator_id = showItem(categoryName,itemName).user_id
-    if login_session['user_id'] != creator_id
-    it = showItem(categoryName,itemName)
+    creator = showItem(categoryName,itemName).user
+    if login_session['user_id'] != creator.id:
+        flash('you are not authorized to edit this item','danger')
+        return redirect(url_for('categoryItem',categoryName=categoryName,itemName=itemName))
     if request.method == 'POST':
         try:
             it = editItem(it,request.form["name"],request.form["description"],request.form["categoryName"])
@@ -225,13 +234,25 @@ def categoryItemEdit(categoryName,itemName):
 #@auth.login_required
 def categoryItemDelete(categoryName,itemName):
     if not isLoggedIn():
+        flash('login is required to create a new category','warning')
         return redirect('/login')
+    creator = showItem(categoryName,itemName).user
+    if login_session['user_id'] != creator.id:
+        flash('you are not authorized to delete this item','danger')
+        return redirect(url_for('categoryItem',categoryName=categoryName,itemName=itemName))
     it = showItem(categoryName,itemName)
     return render_template('category_item_delete.html',item=it)
 
 @app.route('/catalog/<categoryName>/<itemName>/deleteConfirm')
 #@auth.login_required
 def categoryItemDeleteConfirm(categoryName,itemName):
+    if not isLoggedIn():
+        flash('login is required to create a new category','warning')
+        return redirect('/login')
+    creator = showItem(categoryName,itemName).user
+    if login_session['user_id'] != creator.id:
+        flash('you are not authorized to edit this item','danger')
+        return redirect(url_for('categoryItem',categoryName=categoryName,itemName=itemName))
         it = showItem(categoryName,itemName)
         try:
             it = deleteItem(it)
@@ -301,93 +322,90 @@ def oauthDisconnect():
         response.headers['Content-Type'] = 'application/json'
         return response
 
-@app.route('/oauth/<provider>', methods=['POST'])
-def oauthConnect(provider):
+@app.route('/oauth/<provider>', methods = ['POST'])
+def login(provider):
+    #STEP 1 - Parse the auth code
+    auth_code = request.json.get('auth_code')
+    print "Step 1 - Complete, received auth code %s" % auth_code
     if provider == 'google':
-        # Validate state token
-        if request.args.get('state') != login_session['state']:
-            response = make_response(json.dumps('Invalid state parameter.'), 401)
-            response.headers['Content-Type'] = 'application/json'
-            return response
-        # Obtain authorization code
-        code = request.data
-
+        #STEP 2 - Exchange for a token
         try:
             # Upgrade the authorization code into a credentials object
             oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
             oauth_flow.redirect_uri = 'postmessage'
-            credentials = oauth_flow.step2_exchange(code)
+            credentials = oauth_flow.step2_exchange(auth_code)
         except FlowExchangeError:
-            response = make_response(
-                json.dumps('Failed to upgrade the authorization code.'), 401)
+            response = make_response(json.dumps('Failed to upgrade the authorization code.'), 401)
             response.headers['Content-Type'] = 'application/json'
             return response
 
         # Check that the access token is valid.
         access_token = credentials.access_token
-        url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
-               % access_token)
+        url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s' % access_token)
         h = httplib2.Http()
         result = json.loads(h.request(url, 'GET')[1])
         # If there was an error in the access token info, abort.
         if result.get('error') is not None:
             response = make_response(json.dumps(result.get('error')), 500)
             response.headers['Content-Type'] = 'application/json'
-            return response
 
-        # Verify that the access token is used for the intended user.
-        gplus_id = credentials.id_token['sub']
-        if result['user_id'] != gplus_id:
-            response = make_response(
-                json.dumps("Token's user ID doesn't match given user ID."), 401)
-            response.headers['Content-Type'] = 'application/json'
-            return response
+        # # Verify that the access token is used for the intended user.
+        # gplus_id = credentials.id_token['sub']
+        # if result['user_id'] != gplus_id:
+        #     response = make_response(json.dumps("Token's user ID doesn't match given user ID."), 401)
+        #     response.headers['Content-Type'] = 'application/json'
+        #     return response
 
-        # Verify that the access token is valid for this app.
-        if result['issued_to'] != CLIENT_ID:
-            response = make_response(
-                json.dumps("Token's client ID does not match app's."), 401)
-            print "Token's client ID does not match app's."
-            response.headers['Content-Type'] = 'application/json'
-            return response
+        # # Verify that the access token is valid for this app.
+        # if result['issued_to'] != CLIENT_ID:
+        #     response = make_response(json.dumps("Token's client ID does not match app's."), 401)
+        #     response.headers['Content-Type'] = 'application/json'
+        #     return response
 
-        stored_access_token = login_session.get('access_token')
-        stored_gplus_id = login_session.get('gplus_id')
-        if stored_access_token is not None and gplus_id == stored_gplus_id:
-            response = make_response(json.dumps('Current user is already connected.'),
-                                     200)
-            response.headers['Content-Type'] = 'application/json'
-            return response
+        # stored_credentials = login_session.get('credentials')
+        # stored_gplus_id = login_session.get('gplus_id')
+        # if stored_credentials is not None and gplus_id == stored_gplus_id:
+        #     response = make_response(json.dumps('Current user is already connected.'), 200)
+        #     response.headers['Content-Type'] = 'application/json'
+        #     return response
+        print "Step 2 Complete! Access Token : %s " % credentials.access_token
 
-        # Store the access token in the session for later use.
-        login_session['access_token'] = credentials.access_token
-        login_session['gplus_id'] = gplus_id
+        #STEP 3 - Find User or make a new one
 
-        # Get user info
-        userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
-        params = {'access_token': credentials.access_token, 'alt': 'json'}
+        #Get user info
+        h = httplib2.Http()
+        userinfo_url =  "https://www.googleapis.com/oauth2/v1/userinfo"
+        params = {'access_token': credentials.access_token, 'alt':'json'}
         answer = requests.get(userinfo_url, params=params)
 
         data = answer.json()
 
-        login_session['username'] = data['name']
-        login_session['picture'] = data['picture']
-        login_session['email'] = data['email']
+        name = data['name']
+        picture = data['picture']
+        email = data['email']
 
-        output = ''
-        output += '<h1>Welcome, '
-        output += login_session['username']
-        output += '!</h1>'
-        output += '<img src="'
-        output += login_session['picture']
-        output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
-        user_id = getUserID(login_session['email'])
-        if not user_id:
-            user_id = createUser(login_session)
-        login_session['user_id'] = user_id
-        flash("you are now logged in as %s" % login_session['username'])
-        print "done!"
-        return output
+
+
+        #see if user exists, if it doesn't make a new one
+        user = session.query(User).filter_by(email=email).first()
+        if not user:
+            user = User(username = name, picture = picture, email = email)
+            session.add(user)
+            session.commit()
+
+
+
+        #STEP 4 - Make token
+        token = user.generate_auth_token(600)
+
+
+
+        #STEP 5 - Send back token to the client
+        return jsonify({'token': token.decode('ascii')})
+
+        #return jsonify({'token': token.decode('ascii'), 'duration': 600})
+    else:
+        return 'Unrecoginized Provider'
 
 @app.route('/api/users/<int:id>')
 def get_user(id):
